patch to AdMob to fix internal PR review.
